requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
schedule==1.2.1
requests==2.31.0 